refactor(constant): 관심사에 맞게 상수 파일 분리
